move App.jsx to separate file
